output: make tls store flag and frame window conversions explicit
~SSL_TLS_LOG_PEM is a negative int applied to the uint32_t
cert_log_flag, and the frame logging window subtracts a signed offset
from the unsigned stream progress with an intentionally negative
result. Both tripped -fsanitize=implicit-conversion in SV runs: cast
them explicitly, no behaviour change.

+++ src/output-json-frame.c
@@ -353,7 +353,7 @@ static int FrameJson(ThreadVars *tv, JsonFrameLogThread *aft, const Packet *p)
                 continue;
 
             int64_t abs_offset = (int64_t)frame->offset + (int64_t)STREAM_BASE_OFFSET(stream);
-            int64_t win = STREAM_APP_PROGRESS(stream) - abs_offset;
+            int64_t win = (int64_t)STREAM_APP_PROGRESS(stream) - abs_offset;
 
             /* skip frame if threshold not yet reached, esp if frame length is
              * still unknown. */
